Add a script to find links in wikipedia
The script get Wiki link and save it in the file. Take the next random
link from the site, check if the link has not been downloaded once,
save it and get Wiki links from the new site. Take the next random
link ... . The program works in the endless loop.